refactor: move mock generation to test files
